WICKET-1871
git-svn-id: https://svn.apache.org/repos/asf/wicket/branches/wicket-1.3.x@711792 13f79535-47bb-0310-9956-ffa450edef68

+++ jdk-1.4/wicket/src/main/java/org/apache/wicket/util/string/Strings.java
@@ -903,6 +903,14 @@ public final class Strings
 			{
 				ixEnd = path.length();
 			}
+
+			if (ixEnd < ixSemiColon)
+			{
+				// ? is before ; - no jsessionid in the url
+				return url.toString();
+			}
+
+
 			path.delete(ixSemiColon, ixEnd);
 		}
 		return path.toString();
